requirements.txt
streamlit 
ollama 
pytesseract
Pillow